Allow for password reset as SystemAdmin

+++ MemberTrack.Client/src/view-models/users.view-model.ts
@@ -1,28 +1,46 @@
 import { autoinject } from 'aurelia-framework';
+import { EventAggregator } from 'aurelia-event-aggregator';
 
 import { PromptDialogViewModel } from '../view-models/prompt-dialog.view-model';
 import { UserDialogViewModel } from '../view-models/user-dialog.view-model';
 import { UserService } from '../services/user.service';
+import { AuthService } from '../services/auth.service';
 import { BaseViewModel } from '../core/base-view-model';
 import { UserDto, DtoHelper } from '../core/dtos';
-import { PromptEvent, UserEvent } from '../core/custom-events';
+import { PromptEvent, UserEvent, SnackbarEvent } from '../core/custom-events';
+
+const SystemAdminRole = 4;
 
 @autoinject
 export class UsersViewModel extends BaseViewModel {
     private userService: UserService;
+    private authService: AuthService;
+    private eventAggregator: EventAggregator;
     private dtoHelper: DtoHelper;
     public users: UserDto[];
     public promptDialogVm: PromptDialogViewModel;
     public userDialogVm: UserDialogViewModel;
+    private deleteMode: boolean;
+    public canResetPassword: boolean;
 
-    constructor(userService: UserService, dtoHelper: DtoHelper) {
+    constructor(userService: UserService, authService: AuthService, eventAggregator: EventAggregator, dtoHelper: DtoHelper) {
         super('users');
         this.userService = userService;
+        this.authService = authService;
+        this.eventAggregator = eventAggregator;
         this.dtoHelper = dtoHelper;
         this.users = [];
+        this.deleteMode = true;
     }
 
     public attached(): void {
+        this.authService.getContextUser().then(user => {
+            if (!user) {
+                return;
+            }
+            this.canResetPassword = user.role === SystemAdminRole;
+        });
+
         this.userService.getAll().then(dtos => {
             if (!dtos) {
                 return;
@@ -43,18 +61,35 @@ export class UsersViewModel extends BaseViewModel {
     }
 
     public displayPromptDialog(dto: UserDto): void {
+        this.deleteMode = true;
         this.promptDialogVm.show('Delete', 'Are you sure want to delete this user?', dto.displayName, dto.id);
+    }
+
+    public displayResetDialog(dto: UserDto): void {
+        this.deleteMode = false;
+        this.promptDialogVm.show('Reset Password for', 'Are you sure want to reset this user\'s password?', dto.displayName, dto.id);
     }
 
     public dismissPromptDialog(e: CustomEvent): void {
         let event = e.detail.args as PromptEvent;
 
-        this.userService.remove(event.data).then(ok => {
+        if (this.deleteMode) {
+            this.userService.remove(event.data).then(ok => {
+                if (!ok) {
+                    return;
+                }
+                this.dtoHelper.remove(this.users, event.data);
+                this.updateLoadingText(this.users.length);
+            });
+
+            return;
+        }
+
+        this.userService.resetPassword(event.data).then(ok => {
             if (!ok) {
                 return;
             }
-            this.dtoHelper.remove(this.users, event.data);
-            this.updateLoadingText(this.users.length);
+            this.eventAggregator.publish(new SnackbarEvent('Password has been reset!'));
         });
     }
 }
